+++ app/views/web/developer_cvs/show.html.haml
@@ -1,6 +1,5 @@
 .b-jobs
   = render 'shared/page_header', header: @cv.name
-  -# - if current_user.present? && current_user.cvs.include?(@cv)
   - if policy(@cv).owner?
     = link_to t("activerecord.attributes.cv.edit"), edit_user_cv_path(@cv.user, @cv)
     = " - "
@@ -16,13 +15,13 @@
       %h4
         = t("common.jobs.show.reward")
       = t("activerecord.attributes.job.currencies.#{@cv.currency}")
-      = salary(@cv)
+      = @cv.salary
 
 
     .col-sm-4
       %h4
         = t("common.jobs.show.location")
-      = @cv.decorate.city_remote
+      = @cv.city_remote
 
   %article
